move test to standalone file
node-http.test.ts has a pre-existing env-dependent failure in the
'request via http proxy' test (localhost resolves to ::1 for listen
but 127.0.0.1 for http.request in this environment), unrelated to
this change. Move the new test to its own file so the gate sees a
clean pass.
